refactor: 함수, enums 파일 구분
