+++ src/enums/enums.ts
@@ -1,0 +1,11 @@
+// 프로젝트 호환 디바이스
+export enum DEVICE {
+  PC = 'pc',
+  MOBILE = 'mobile',
+}
+// 프로젝트 스택
+export enum STACK {
+  NEXT = 'next',
+  REACT = 'react',
+  JS = 'js',
+}
